requirements.txt
amqp==2.1.4
beautifulsoup4==4.5.3
billiard==3.5.0.2
celery==4.0.2
Django==1.10.5
django-extra-views==0.7.1
django-filter==0.15.3
django-fsm==2.4.0
django-material==0.8.0
django-viewflow-pro==0.10.10
django-webtest==1.8.0
funcsigs==1.0.2
kombu==4.0.2
mock==2.0.0
pbr==1.10.0
pytz==2016.10
singledispatch==3.4.0.3
six==1.10.0
vine==1.1.3
waitress==1.0.1
WebOb==1.7.0
WebTest==2.0.24
